first attempt to oscarify the code
- changed the code to use Oscar objects where possible;
  now explicit GAP and Singular calls occur only in low level functions;
  polymake is still called "on the surface", this may change in a further step
- updated to Oscar 0.5.0, adjusted the output accordingly
- adjusted the Jupyter notebook to the new situation

The old code, which can be seen as a documentation how to use the
interfaces to the Oscar cornerstones directly, is still available,
via a submodule `GITFans.OLD`;
also the old notebook is still there.
