+++ example_interfaces.jl
@@ -14,6 +14,17 @@
 ##  $\overline{M}_{0,6}$](https://arxiv.org/abs/1603.09241)
 ##  (To appear in Math. Comp. 2020.)
 
+
+############################################################################
+##  In this session, we use the functions from the Julia module `GITFans.OLD`,
+##  from the original implementation of the code which focuses on the
+##  interfaces from Julia to GAP, polymake, and Singular.
+##  Thus the functions from these systems are called explicitly.
+##
+##  There is a newer version `example_oscar.jl`, where only Oscar functions
+##  are called, delegations to other systems (where necessary) happen
+##  at lower levels there.
+
 using Oscar
 using GITFans
 
@@ -44,14 +55,11 @@ Q = [
  0  0   0   0   1 ];
 
 # polynomial ring
-n = size(Q, 1)
-Qt, T = PolynomialRing(QQ, :T => 1:n)
-D = free_abelian_group(size(Q,2))
-w = [D(Q[i, :]) for i = 1:n]
-R = grade(Qt, w)
+R, T = Singular.PolynomialRing(Singular.QQ,
+           ["x" * string(i) for i in 1:size(Q, 1)])
 
 # generators for the ideal
-a = ideal(R, [
+a = Singular.Ideal(R, [
     T[5]*T[10] - T[6]*T[9] + T[7]*T[8],
     T[1]*T[9]  - T[2]*T[7] + T[4]*T[5],
     T[1]*T[8]  - T[2]*T[6] + T[3]*T[5],
@@ -64,19 +72,19 @@ a = ideal(R, [
 ##  We observe that there is an $S_5$-symmetry
 ##  for the $H\cong (\KK^*)^5$-action on $V(\aa)$ where the symmetry group
 ##  $S_5\cong G\subseteq S_{10}$ is generated by
-##  (2,3)(5,6)(9,10), (1,5,9,10,3)(2,7,8,4,6).
+##  (2,3)(5,6)(9,10), (1,5,9,10,3)(2,7,8,4,6)
+##  and represented as a group in GAP.
 
-perms_list = [ [1,3,2,4,6,5,7,8,10,9], [5,7,1,6,9,2,8,4,10,3] ];
-sym10 = symmetric_group(n);
-G, emb = sub([sym10(x) for x in perms_list]...);
-G
+perms_list = [[1,3,2,4,6,5,7,8,10,9], [5,7,1,6,9,2,8,4,10,3]];
+perms = [GAP.Globals.PermList(GAP.julia_to_gap(i)) for i in perms_list];
+G = GAP.Globals.Group(GAP.julia_to_gap(perms))
 
 
 ############################################################################
 ##  We now compute the GIT-fan, represented as a fan in polymake,
 ##  using Gröbner bases from Singular:
 
-fanobj = GITFans.git_fan(a, Q, G)
+fanobj = GITFans.OLD.git_fan(a, Q, G)
 
 
 ############################################################################
@@ -90,22 +98,23 @@ fanobj.F_VECTOR
 ##
 ##  We compute the orbit cones as projection of a-faces and partition
 ##  the set of orbit cones into orbits under the symmetry group action.
-##  We also return the action on the orbit cones in terms of homomorphisms.
+##  We also return the action on the orbit cones in terms of
+##  GAP homomorphisms.
 
-oc = GITFans.orbit_cone_orbits_and_action(a, Q, G);
+oc = GITFans.OLD.orbit_cone_orbits_and_action(a, Q, G);
 
 
 ############################################################################
 ##  Lengths of the orbit cone orbits:
 
-map( length, oc[ "orbit_list" ] )
+map(length, oc["orbit_list"])
 
 
 ############################################################################
 ##  Action of the symmetry group
 ##  permuting the elements of the first orbit cone orbit:
 
-oc[ "homs" ][1]
+oc["homs"][1]
 
 
 ############################################################################
@@ -115,14 +124,14 @@ oc[ "homs" ][1]
 ##  where the GIT-cones are described via hashes encoding the cones
 ##  as intersections of orbits cones.
 ##  The data structure also contains the group action on the hashes
-##  encoded as homomorphisms.
+##  encoded as GAP homomorphisms.
 ##
 ##  The algorithm is based on a fan traversal.
 ##
 ##  The function also returns the incidence relation of the orbits of
 ##  GIT-cones.
 
-(hash_list, edges) = GITFans.fan_traversal(oc);
+(hash_list, edges) = GITFans.OLD.fan_traversal(oc);
 
 
 ############################################################################
@@ -150,7 +159,7 @@ Polymake.graph.visual(intergraph)
 ##  We translate the descriptions of the six maximal cones back
 ##  to cone objects and expand their G-orbits.
 
-expanded = GITFans.orbits_of_maximal_GIT_cones(oc, hash_list);
+expanded = GITFans.OLD.orbits_of_maximal_GIT_cones(oc, hash_list);
 orbit_lengths = map(length, expanded)
 
 
@@ -173,7 +182,7 @@ sum(orbit_lengths)
 ############################################################################
 ##  We create the GIT-Fan represented by a fan object in polymake:
 
-fanobj = GITFans.hashes_to_polyhedral_fan(oc, hash_list)
+fanobj = GITFans.OLD.hashes_to_polyhedral_fan(oc, hash_list)
 
 
 ############################################################################
